user_service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.20"
uvicorn = {extras = ["standard"], version = "^0.30.1"}

passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiokafka = "^0.11.0"
asyncio = "^3.4.3"
starlette = "^0.37.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"
sqlalchemy = "^2.0.31"
bcrypt = "^4.2.0"

psycopg = "^3.2.1"
fastapi = "0.111.1"
pydantic = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
